requirements.txt
asgiref==3.4.1
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.0.4
cmake==3.26.3
datetime-truncate==1.1.1
Django==3.2.18
django-admin-tools==0.9.2
django-bower==5.2.0
django-countries==7.5.1
django-extensions==3.2.1
django-jazzmin==2.6.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-webcampicture==0.1.9
dlib
dlib-bin
face-recognition
face-recognition-models
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy
opencv-python
Pillow==8.4.0
pipdeptree==2.2.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2023.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.4
stripe==2.63.0
typing_extensions==4.1.1
Unidecode==1.3.6
urllib3==1.26.15
paypalrestsdk
stripe
paystackapi
paystack